Crate infinity_compiler

Compiler between IE game files and a SQLite database.

Modules

database

Access to the SQL side of the database.

gamefiles

Access to the KEY/BIF side of the database.

gamestrings

Access to game strings in database.

lua_api

Loads mod-supplied Lua files.

pack

I/O to binary format. See the Pack trait.

prelude 🔒

The set of symbols we want accessible from everywhere in the crate.

resources

Generic interface for game resources. Definition of specific resources goes to restypes mod.

restypes 🔒

Full definition of all structures representing game resources.

schemas

Inner description of SQL table.

sql_rows

Basic types for interaction with SQL and binary files. The main entry points for this module are the traits Pack and SqlRow.

staticstrings

Fixed-length string used as resource identifier.

toolbox

Generic useful functions for user interaction.

Structs

RuntimeOptions 🔒

The runtime options parser for clap crate.

Enums

Command 🔒

Subcommands passed to the executable.

Functions

command_restore 🔒

Restores all backed up files.

command_save_diff 🔒

command_save_full 🔒

Saves game resources to filesystem.

command_show 🔒

main 🔒

save_resources 🔒

Saves all modified game strings and resources to current directory.

type_of 🔒