Struct infinity_compiler::database::GameDB

pub struct GameDB(Connection);

A trivial wrapper on rusqlite::Connection; mainly used for standardizing log messages.

Tuple Fields
0: Connection
Implementations

impl GameDB

pub fn transaction<T>(
&mut self,
f: impl FnMut(&GameTransaction<'_>) -> Result<T>
) -> Result<T>

Wraps a closure inside a game transaction.

The transaction aborts if the closure returns an Err variant, and commits if it returns an Ok variant.

pub fn open(p: impl AsRef<Path>) -> Result<Self>

Opens an existing game database.

(Just like Connection::open but with logging).

pub fn create(db_file: impl AsRef<Path>, game: &GameIndex) -> Result<Self>

Creates and initializes the game database.

This creates all relevant tables and views in the database (but does not fill them).

There are global tables, per-resource tables, and per-language tables. The per-resource tables are described in the crate::schemas::Schema::create_tables_and_views function. The per-language tables are described in Self::create_language_tables. The global tables are the following:

global: a single-row table holding the global variables:

component: current mod component being installed;

strref_count: number of strings in "dialog.tlk" (34000);

resref_orig: list of all original game resrefs — this is used for de-namespacing (all resrefs in this list translate to themselves);

resref_dict: translations of resrefs (from namespaced to 8 bytes);

strref_dict: translations of strrefs (from string key to 4-bytes int);

new_strings: a view of all strrefs currently introduced by mods. This gets compiled into strref_dict as part of save.

fn create_language_tables(&mut self, game: &GameIndex) -> Result<()>

Creates (but does not fill) all per-language tables.

The per-language tables are as follows:

translations_XX: user-filled table containing string translations

strings_XX: output to fill game strings (last column is a bit indicating whether this is untranslated, in which case we need to remove all {?..} marks)

pub fn load(&mut self, game: &GameIndex) -> Result<()>

Fills the database from game files.

The database must have been already created and initialized (with empty tables).