Trait infinity_compiler::database::DbInterface

pub trait DbInterface {
    fn execute(
        &self,
        s: impl AsRef<str>,
        params: impl Params + Debug
    ) -> Result<usize>;
    fn prepare(&self, s: impl AsRef<str>) -> Result<Statement<'_>>;
    fn last_insert_rowid(&self) -> i64;
    fn batch(&self, s: impl AsRef<str>) -> Result<()>;

    fn exec(&self, s: impl AsRef<str>) -> Result<usize> { ... }
    fn query_row<T>(
        &self,
        s: impl AsRef<str>,
        params: impl Params + Debug,
        f: impl FnOnce(&Row<'_>) -> Result<T>
    ) -> Result<T> { ... }
    fn query_row_and_then<T>(
        &self,
        s: impl AsRef<str>,
        params: impl Params + Debug,
        f: impl FnOnce(&Row<'_>) -> Result<T>
    ) -> Result<T> { ... }
    fn translate_strrefs(&self) -> Result<()> { ... }
    fn translate_resrefs(&self) -> Result<()> { ... }
    fn clear_orphan_resources(&self) -> Result<()> { ... }
    fn unmark_dirty_resources(&self) -> Result<()> { ... }
}

A trait for trivial wrappers of rusqlite::Connection.

Required Methods

fn execute(
&self,
s: impl AsRef<str>,
params: impl Params + Debug
) -> Result<usize>

Executes a statement with arguments (and logging).

fn prepare(&self, s: impl AsRef<str>) -> Result<Statement<'_>>

Prepares a statement (and logs it).

fn last_insert_rowid(&self) -> i64

fn batch(&self, s: impl AsRef<str>) -> Result<()>

Executes a batch of statements (with logging).

Provided Methods

fn exec(&self, s: impl AsRef<str>) -> Result<usize>

Executes a statement without arguments (but with logging).

fn query_row<T>(
&self,
s: impl AsRef<str>,
params: impl Params + Debug,
f: impl FnOnce(&Row<'_>) -> Result<T>
) -> Result<T>

Prepares and executes a statement returning a single row.

fn query_row_and_then<T>(
&self,
s: impl AsRef<str>,
params: impl Params + Debug,
f: impl FnOnce(&Row<'_>) -> Result<T>
) -> Result<T>

fn translate_strrefs(&self) -> Result<()>

Before saving: fills the strref translation tables.

fn translate_resrefs(&self) -> Result<()>

Before saving: fills the resref translation table.

fn clear_orphan_resources(&self) -> Result<()>

Deletes from current directory all resources marked as ‘orphan’ in the database. This also clears the list of orphan resources.

fn unmark_dirty_resources(&self) -> Result<()>

Cleans the dirty bit from all resources after saving.