Trait infinity_compiler::pack::Pack

pub trait Pack: Sized {
    fn unpack(f: &mut impl Read) -> Result<Self>;

    fn pack(&self, _f: &mut impl Write) -> Result<()> { ... }
    fn vecunpack(f: &mut impl Read, n: usize) -> Result<Vec<Self>> { ... }
    fn read_bytes(f: &mut impl Read, n: usize) -> Result<Vec<u8>> { ... }
    fn unpack_header(f: &mut impl Read, hdr: &str) -> Result<()> { ... }
}

A type which may be packed to binary.

This trait is implemented by the derive macro. The macro may be applied to both named and unnamed structs, as well as (in some limited cases) for parametrized structs (in which case it will require that all fields implement Pack).

Required Methods

fn unpack(f: &mut impl Read) -> Result<Self>

Reads this object from a binary source.

Provided Methods

fn pack(&self, _f: &mut impl Write) -> Result<()>

Writes this objects to a binary sink.

fn vecunpack(f: &mut impl Read, n: usize) -> Result<Vec<Self>>

Reads a vector of objects (of known size) from a binary source.

fn read_bytes(f: &mut impl Read, n: usize) -> Result<Vec<u8>>

Helper function used to check forced headers.

fn unpack_header(f: &mut impl Read, hdr: &str) -> Result<()>

function which checks that a header is correct.