Module infinity_compiler::resources

Full definition of all structures representing game resources.

This crate makes heavy use of the Pack and Resource0 derive macros to automatically interface with game files and the SQL database.

Most of the Lua-side definition of resource schemas is also derived from here. The exception is the relation between resources and their subresources. For now, this is inserted by hand by the init.lua file.

The tables share the following common structure:

Payload fields: these carry actual game information and and are defined in game resources. These always come as the first columns so that select statements can load resource fields from predictible column indices.

Context fields: these identify in which game resource, and where, this information is located. The first context field is always the resref for the main resource. These fields also carry a unique constraint.

For subresources: a rowid alias (integer primary key). (Top-level resources don’t need this: they have the resref as their primary key).

The integer primary key is needed even for subresources: namely, we want these to have a persistent identifier for joins with the edit tables.

Structs

AllResources

A struct holding one field for each game resource type.

Item

A game item, corresponding to a .itm file.

ItemAbility

An ability inside a .itm file.

ItemEffect

An effect inside a .itm file (either global or in an ability).

Traits

Resource

One of the resource tables stored in the database.

ToplevelResource

A top-level resource (i.e. saved to its own file in the override directory).

ToplevelResourceData

Functions

all_schemas

restype_from_extension

When provided with a file extension, return the corresponding Restype.