Struct infinity_compiler::resources::Tree

pub struct Tree<T: ByName> {
    pub content: T::In,
    pub branches: T,
}

A hierarchical collection of similar objects, organized in parallel with the resources and subresources in the database.

Fields
content: T::Inbranches: T