Trait infinity_compiler::resources::Recurse

pub trait Recurse<Y>: DerefMut {
    type To;

    fn recurse<'n, S, E, F>(
        &self,
        f: F,
        name: &'n str,
        state: Option<&S>
    ) -> Result<Self::To, E>
    where
        F: Fn(&Self::Target, &'n str, Option<&S>) -> Result<(S, Y), E>;

    fn try_map<E, F>(&self, f: F) -> Result<Self::To, E>
    where
        F: Fn(&Self::Target) -> Result<Y, E>,
    { ... }
    fn map<F>(&self, f: F) -> Self::To
    where
        F: Fn(&Self::Target) -> Y,
    { ... }
}

Values organized along a tree matching the structure of resources in the database.

This is a crude version of a higher-kinded type; the value is wrapped in a DerefMut (the type is <Self as Deref>::Target).

This type is implemented by the Resource derive macro. It has sub-fields matching the sub-resources for this resource, and the implementation for the Self::recurse method iterate over these fields.

Required Associated Types

type To

Result of replacing current parameter by Y.

Required Methods

fn recurse<'n, S, E, F>(
&self,
f: F,
name: &'n str,
state: Option<&S>
) -> Result<Self::To, E>where
F: Fn(&Self::Target, &'n str, Option<&S>) -> Result<(S, Y), E>,

Recursively traverse the tree from its root, applying the closure on each element. Used by traverse and map.

For each node in the tree, the closure f is invoked with:

the content of the node,

the name of this node (as a &'static str),

the state computed for the parent (as an Option), It should return a tuple of:

the new state value passed to children,

the value computed for this node.

Provided Methods

fn try_map<E, F>(&self, f: F) -> Result<Self::To, E>where
F: Fn(&Self::Target) -> Result<Y, E>,

Apply a closure to each element of the tree; fallible version.

fn map<F>(&self, f: F) -> Self::Towhere
F: Fn(&Self::Target) -> Y,

Apply a closure to each element of the tree; infallible version.