Trait infinity_compiler::resources::Resource

pub trait Resource: SqlRow {
    type FieldNode;
    type Primary: ToSql + FromSql + Display + Copy;
    type StatementNode<'a>: DerefMut<Target = Statement<'a>>;

    const FIELDS_NODE: Self::FieldNode;

    fn insert_subresources(
        &self,
        db: &Connection,
        node: &mut Self::StatementNode<'_>,
        primary: impl ToSql + Copy
    ) -> Result<()>;

    fn insert_as_topresource(
        &self,
        db: &Connection,
        node: &mut Self::StatementNode<'_>,
        resref: Self::Primary
    ) -> Result<usize> { ... }
    fn insert_as_subresource(
        &self,
        db: &Connection,
        node: &mut Self::StatementNode<'_>,
        parent: impl ToSql,
        position: usize
    ) -> Result<()> { ... }
    fn iter_rows<'a>(
        s: &'a mut Statement<'_>,
        params: impl Params
    ) -> Result<TypedRows<'a, (Self::Primary, Self)>> { ... }
    fn collect_rows(
        node: &mut Self::StatementNode<'_>,
        params: impl Params
    ) -> Result<Vec<Self>> { ... }
}

The trait shared by all game resources. All the required functions are produced by the Resource derive macro from the list of sub-resources for this resource.

Required Associated Types

type FieldNode

Always FooNode<Fields>.

type Primary: ToSql + FromSql + Display + Copy

The type of the primary key for this resource: either Resref for top-level or i64 for sub-resources.

type StatementNode<'a>: DerefMut<Target = Statement<'a>>

Always FooNode<Statement<'a>>.

Required Associated Constants

const FIELDS_NODE: Self::FieldNode

The node in the tree holding the fields description for this resource.

Required Methods

fn insert_subresources(
&self,
db: &Connection,
node: &mut Self::StatementNode<'_>,
primary: impl ToSql + Copy
) -> Result<()>

Given an object inserted into the database, insert its sub-resources. Generated by macro and mutually recursive with (subresources)::insert_as_subresource.

Provided Methods

fn insert_as_topresource(
&self,
db: &Connection,
node: &mut Self::StatementNode<'_>,
resref: Self::Primary
) -> Result<usize>

Inserts a top-level resource. This differs from subresources in that the primary key (resref) is known.

This is the entry point called when populating the database; this function is not called recursively.

fn insert_as_subresource(
&self,
db: &Connection,
node: &mut Self::StatementNode<'_>,
parent: impl ToSql,
position: usize
) -> Result<()>

Inserts a sub-resource.

This differs from a top-level resource in that the primary key (rowid) is not known yet; instead it is derived from last_insert_rowid.

This function is mutually recursive with Self::insert_subresources (derived from macro).

fn iter_rows<'a>(
s: &'a mut Statement<'_>,
params: impl Params
) -> Result<TypedRows<'a, (Self::Primary, Self)>>

Converts a statement into an iterator returning structured results.

fn collect_rows(
node: &mut Self::StatementNode<'_>,
params: impl Params
) -> Result<Vec<Self>>

Collects all rows returned by a statement into a vector.