Trait infinity_compiler::resources::ResourceTree

pub trait ResourceTree: SqlRow {
    type FieldsTree;
    type StatementForest<'a>: 'a + ByName<In = Statement<'a>>;
    type Primary: ToSql + FromSql + Display + Copy;

    const FIELDS_TREE: Self::FieldsTree;

    fn insert_subresources(
        &self,
        db: &Connection,
        branches: &mut Self::StatementForest<'_>,
        primary: impl ToSql + Copy
    ) -> Result<()>;
    fn select_subresources(
        &mut self,
        branches: &mut Self::StatementForest<'_>,
        primary: Self::Primary
    ) -> Result<()>;

    fn insert_as_topresource(
        &self,
        db: &Connection,
        tree: &mut Tree<Self::StatementForest<'_>>,
        resref: Self::Primary
    ) -> Result<usize> { ... }
    fn insert_as_subresource(
        &self,
        db: &Connection,
        tree: &mut Tree<Self::StatementForest<'_>>,
        parent: impl ToSql,
        position: usize
    ) -> Result<()> { ... }
    fn read_rows<'a: 'r, 'b: 'r, 'r>(
        tree: &'a mut Tree<Self::StatementForest<'b>>,
        params: impl Params
    ) -> Result<RecursiveRows<'r, 'a, 'b, Self>> { ... }
    fn collect_all(
        tree: &mut Tree<Self::StatementForest<'_>>,
        params: impl Params
    ) -> Result<Vec<Self>> { ... }
}

Required Associated Types

type FieldsTree

Always Tree<FooForest<Fields>>.

type StatementForest<'a>: 'a + ByName<In = Statement<'a>>

SQL statements in the database, linked to all the resources tables.

type Primary: ToSql + FromSql + Display + Copy

The type of the primary key for this resource: either Resref for top-level or i64 for sub-resources.

Required Associated Constants

const FIELDS_TREE: Self::FieldsTree

The tree holding the fields description for this resource and all sub-resources.

Required Methods

fn insert_subresources(
&self,
db: &Connection,
branches: &mut Self::StatementForest<'_>,
primary: impl ToSql + Copy
) -> Result<()>

Given an object inserted into the database, insert its sub-resources. Generated by macro. and mutually recursive with (subresources)::insert_as_subresource.

fn select_subresources(
&mut self,
branches: &mut Self::StatementForest<'_>,
primary: Self::Primary
) -> Result<()>

Complete a partially-initialized object from the db by reading all its sub-resource fields. Generated by macro.

Provided Methods

fn insert_as_topresource(
&self,
db: &Connection,
tree: &mut Tree<Self::StatementForest<'_>>,
resref: Self::Primary
) -> Result<usize>

Inserts a top-level resource. This differs from subresources in that the primary key (resref) is known.

This is the entry point called when populating the database; this function is not called recursively.

fn insert_as_subresource(
&self,
db: &Connection,
tree: &mut Tree<Self::StatementForest<'_>>,
parent: impl ToSql,
position: usize
) -> Result<()>

Inserts a sub-resource.

This differs from a top-level resource in that the primary key (rowid) is not known yet; instead it is derived from last_insert_rowid.

This function is mutually recursive with Self::insert_subresources (derived from macro).

fn read_rows<'a: 'r, 'b: 'r, 'r>(
tree: &'a mut Tree<Self::StatementForest<'b>>,
params: impl Params
) -> Result<RecursiveRows<'r, 'a, 'b, Self>>

fn collect_all(
tree: &mut Tree<Self::StatementForest<'_>>,
params: impl Params
) -> Result<Vec<Self>>