Trait infinity_compiler::resources::TopResource

pub trait TopResource: SqlRow {
    type Subresources;

    const EXTENSION: &'static str;
    const RESTYPE: Restype;

    fn load(
        tables: &mut AllTables<Statement<'_>>,
        db: &impl DbInterface,
        cursor: impl Read + Seek,
        resref: Resref
    ) -> Result<()>;
    fn save(
        &mut self,
        file: impl Write + Seek + Debug,
        subresources: &Self::Subresources
    ) -> Result<()>;
    fn select_subresources(
        &mut self,
        tables: &mut AllTables<Statement<'_>>,
        resref: Resref
    ) -> Result<Self::Subresources>;

    fn load_from_handle<T: DbInterface>(
        db: &mut DbInserter<'_, T>,
        resref: Resref,
        handle: ResHandle<'_>
    ) -> Result<()> { ... }
    fn save_all_dirty(
        db: &impl DbInterface,
        tables: &mut AllTables<Statement<'_>>,
        name: impl Display
    ) -> Result<usize> { ... }
    fn iter_rows<'a>(
        s: &'a mut Statement<'_>,
        params: impl Params
    ) -> Result<TypedRows<'a, (Resref, Self)>> { ... }
}

One of the resource tables stored in the database.

Methods attached to this trait are those which depend both on

having a full schema (with table name, etc.), and

having access to the attached (Rust) type (for typed inserts, etc.). Methods depending only on the list of columns of the schema and the Rust struct go to SqlRow. Methods depending only on the schema go to Schema. Methods depending only on the list of columns from the schema (i.e. the intersection of both cases) go to crate::schemas::Fields

Required Associated Types

type Subresources

Which subresources do we need to collect from the database.

Required Associated Constants

const EXTENSION: &'static str

The file extension attached to this resource.

const RESTYPE: Restype

The numeric identifier for this resource (e.g. 0x03ed).

Required Methods

fn load(
tables: &mut AllTables<Statement<'_>>,
db: &impl DbInterface,
cursor: impl Read + Seek,
resref: Resref
) -> Result<()>

Loads a resource from filesystem directly into database.

fn save(
&mut self,
file: impl Write + Seek + Debug,
subresources: &Self::Subresources
) -> Result<()>

Saves a resource (from struct + subresources) to filesystem.

fn select_subresources(
&mut self,
tables: &mut AllTables<Statement<'_>>,
resref: Resref
) -> Result<Self::Subresources>

Selects subresources from database.

Provided Methods

fn load_from_handle<T: DbInterface>(
db: &mut DbInserter<'_, T>,
resref: Resref,
handle: ResHandle<'_>
) -> Result<()>

Loads a resource from filesystem to database. Main entry point for the init command.

fn save_all_dirty(
db: &impl DbInterface,
tables: &mut AllTables<Statement<'_>>,
name: impl Display
) -> Result<usize>

Saves all dirty resources to the filesystem. Main entry point for the save command.

fn iter_rows<'a>(
s: &'a mut Statement<'_>,
params: impl Params
) -> Result<TypedRows<'a, (Resref, Self)>>

Iterates over rows returned by this statement, as (Resref, Self).