Trait infinity_compiler::resources::ToplevelResource

pub trait ToplevelResource: NamedTable {
    type Subresources<'a>;

    fn load(
        db: &mut DbInserter<'_>,
        cursor: impl Read + Seek,
        resref: Resref
    ) -> Result<()>;
    fn save(
        &mut self,
        file: impl Write + Seek + Debug,
        subresources: Self::Subresources<'_>
    ) -> Result<()>;
    fn for_each<F: Fn(Resref, &mut Self, Self::Subresources<'_>) -> Result<()>>(
        db: &impl DbInterface,
        condition: impl Display,
        f: F
    ) -> Result<i32>;
    fn show(&self, resref: impl Display, subresources: Self::Subresources<'_>);

    fn save_all(db: &impl DbInterface) -> Result<()> { ... }
    fn show_all(db: &impl DbInterface, resref: impl Display) -> Result<()> { ... }
}

A top-level resource (i.e. saved to its own file in the override directory).

Required Associated Types

type Subresources<'a>

Subresources for this resource (together with linking info.).

This should be a tuple of all subresources, each of them stored in some Vec with elements (subresource, linking-info).

Required Methods

fn load(
db: &mut DbInserter<'_>,
cursor: impl Read + Seek,
resref: Resref
) -> Result<()>

Inserts a single resource in the database.

fn save(
&mut self,
file: impl Write + Seek + Debug,
subresources: Self::Subresources<'_>
) -> Result<()>

Saves a single resource (with its subresources) to game files.

fn for_each<F: Fn(Resref, &mut Self, Self::Subresources<'_>) -> Result<()>>(
db: &impl DbInterface,
condition: impl Display,
f: F
) -> Result<i32>

Applies a closure over all resources together with their subresources.

Iterates a closure over items in the database (with their associated properties) matching the given condition.

This function links items with their sub-resources (e.g. item abilities and effects) and calls the passed closure on the resulting structure.

It returns the number of rows matched.

fn show(&self, resref: impl Display, subresources: Self::Subresources<'_>)

Shows a resource on stdout (grouped with its subresources) in a nice user format.

Provided Methods

fn save_all(db: &impl DbInterface) -> Result<()>

Saves all toplevel resources of this type.

fn show_all(db: &impl DbInterface, resref: impl Display) -> Result<()>

Shows on stdout the resource with given resref, or returns an error.