Struct infinity_compiler::restypes::RootNode

pub struct RootNode<X: Debug> {
    pub items: ItemNode<X>,
    _marker: PhantomData<X>,
}

A Node impl. derived by derive(RecursiveResource).

Fields
items: ItemNode<X>_marker: PhantomData<X>
Implementations

impl RootNode<Schema>

pub fn create_new_strings<T>(&self, f: impl Fn(&str) -> Result<T>) -> Result<()>

Builds the SQL statement creating the new_strings view.

This also builds a few triggers which mark resources as dirty whenever their strrefs are reassigned.

impl<'a, T: Callback<'a> + Debug + 'a> RootNode<T>

fn prepare(db: &'a impl DbInterface) -> Result<Self>where
Self: Sized,

fn execute<'lua>(
&mut self,
lua: &'lua Lua,
args: MultiValue<'lua>
) -> Result<Value<'lua>>

Selects the appropriate individual callback from the first argument (table name) and runs it.

fn install_callback<'scope>(
&'scope mut self,
scope: &Scope<'_, 'scope>,
table: &Table<'scope>,
name: &'scope str
) -> Result<()>

The wrapper installing the callback function in a table.

impl<X: Debug> RootNode<X>

pub fn by_name<'a>(&'a self, target: &str) -> Result<&'a X>

Runtime search in the tree. this would be a bit hard to do with recurse — the lifetimes are a mess, and we want to interrupt search as soon as we find and cut branches with a non-matching name:

fn by_name1<'a>(&'a self, tail: &str) -> Option<&'a X>

pub fn by_name_mut<'a>(&'a mut self, target: &str) -> Result<&'a mut X>

Same, with mutable reference.

fn by_name_mut1<'a>(&'a mut self, tail: &str) -> Option<&'a mut X>