Struct infinity_compiler::schemas::Schema

pub struct Schema {
    pub level: usize,
    pub name: String,
    pub table_type: TableType,
    pub fields: Fields,
}

The full database description of a game resource.

This contains all relevant information to fully define a resource on the SQL side.

In practice there exists exactly one Schema instance per resource, and it is compiled by the SqlRow derive macro.

Fields
level: usize

The nesting level for this resource (0 for top-level resources).

name: String

Name of the SQL table mapped to this data.

table_type: TableType

Position of the schema in the arborescence.

fields: Fields

Description of the SQL table for this data.

Implementations

impl Schema

pub fn is_subresource(&self) -> bool

true iff this is not a top-level resource.

pub fn describe(&self)

Displays a full description of the schema on stdout.

pub fn create_table(&self, name: impl Display, more: impl Display) -> String

Creates a table with this schema and an arbitrary name.

pub fn create_main_view(&self) -> String

Returns the SQL code creating the main view (e.g. "items").

pub fn create_save_view(&self) -> String

Creates the save_xxx view.

pub fn create_tables_and_views<T>(
&self,
f: impl Fn(String) -> Result<T>
) -> Result<()>

Creates all tables and views in the database associated with a given resource.

The tables are as follows:

load_{name}: data read from game files.

add_{name}: new data inserted by mods.

edit_{name}: data modified by mods. The views are:

{name}: the main view on which edits are done by mods (and propagated to add_{name} or edit_{name} by triggers as needed).

save_{name}: the view used to save new or modified values to game files.

pub fn insert_sql(&self, prefix: impl Display) -> String

The SQL code for inserting into a table.

This is used for both initial populating of the database from game files and for simod.insert.

The or parameter allows calling SQL INSERT OR IGNORE; this is used when initializing the database to ignore resources which are superseded by override files.

pub fn append_new_strings_schema(&self, w: &mut impl Write, is_first: &mut bool)

Helper function for generating new_strings view.

pub fn append_new_strings_trigger(&self, w: &mut impl Write)

Helper function for generating new_strings triggers.

pub fn select_dirty_sql(&self) -> String

Returns the SQL statement for saving resources to filesystem.