Trait infinity_compiler::sql_rows::AsParams

pub trait AsParams {
    type Elt<'a>: ToSql
    where
        Self: 'a;
    type Iter<'a>: Iterator<Item = Self::Elt<'a>>
    where
        Self: 'a;

    fn params_iter(&self) -> Self::Iter<'_>;

    fn as_params(&self) -> ParamsFromIter<Self::Iter<'_>> { ... }
}

Converting an object to something implementing [rusqlite::Params].

Since we cannot implement the [rusqlite::Params] trait ourselves, this is the closest we can do.

Required Associated Types

type Elt<'a>: ToSql
where
Self: 'a

Element type of the iterator we use to map to SQL.

type Iter<'a>: Iterator<Item = Self::Elt<'a>>
where
Self: 'a

Iterator we use to map to SQL.

Required Methods

fn params_iter(&self) -> Self::Iter<'_>

Converts this to an iterator.

Provided Methods

fn as_params(&self) -> ParamsFromIter<Self::Iter<'_>>

Converts this to some object implementing [rusqlite::Params].