Trait infinity_compiler::sql_rows::SqlRow

pub trait SqlRow: Sized {
    const FIELDS: SqlRowData;
    const FIELDS9: Fields;

    fn bind_at(&self, s: &mut Statement<'_>, offset: usize) -> Result<()>;
    fn collect_at(s: &Row<'_>, offset: usize) -> Result<Self>;

    fn check_parameter_count(s: &mut Statement<'_>, extra: usize) -> Result<()> { ... }
    fn bind_execute1<P: ToSql>(
        &self,
        stmt: &mut Statement<'_>,
        ctx: P
    ) -> Result<usize> { ... }
    fn bind_execute2<P1: ToSql, P2: ToSql>(
        &self,
        stmt: &mut Statement<'_>,
        c1: P1,
        c2: P2
    ) -> Result<usize> { ... }
}

Structure which can be read from (part of) a SQL row, or bound to (part of) a SQL statement.

Required Associated Constants

const FIELDS: SqlRowData

const FIELDS9: Fields

The description of the columns that this binds to.

Required Methods

fn bind_at(&self, s: &mut Statement<'_>, offset: usize) -> Result<()>

Binds to columns [offset, …] of statement.

fn collect_at(s: &Row<'_>, offset: usize) -> Result<Self>

Reads from columns [offset, … ] of row.

Provided Methods

fn check_parameter_count(s: &mut Statement<'_>, extra: usize) -> Result<()>

Checks that the number of columns is good.

fn bind_execute1<P: ToSql>(
&self,
stmt: &mut Statement<'_>,
ctx: P
) -> Result<usize>

Binds this object with a header of a single parameter, and executes the statement.

fn bind_execute2<P1: ToSql, P2: ToSql>(
&self,
stmt: &mut Statement<'_>,
c1: P1,
c2: P2
) -> Result<usize>

Binds this object with a header of two parameters, and executes the statement.