Trait infinity_compiler::struct_io::SqlMapped

pub trait SqlMapped: Sized {
    const WIDTH: usize;

    fn get_field(&self, i: usize) -> ToSqlOutput<'_>;
    fn from_row_at(r: &Row<'_>, offset: usize) -> Result<Self>;

    fn as_params(&self) -> ParamsFromIter<SqlMappedIterator<'_, Self>> { ... }
}

A struct mapped to a SQL row.

Required Associated Constants

const WIDTH: usize

The number of fields in the SQL row.

Required Methods

fn get_field(&self, i: usize) -> ToSqlOutput<'_>

Returns the i-th field of the resource, wrapped in a [rusqlite::types::ToSlqOutput] for trivial insertion into a SQL statement. NOTE: rusqlite does not allow pushing Result<ToSqlOutput> values back into a [rusqlite::ParamsFromIter] iterator, so we need to unwrap those values. Since we do so from a controlled struct, this should not panic; however, it would be better to be able to propagate errors.

fn from_row_at(r: &Row<'_>, offset: usize) -> Result<Self>

Reads a whole [rusqlite::Row] into a struct, or raises a conversion error.

Provided Methods

fn as_params(&self) -> ParamsFromIter<SqlMappedIterator<'_, Self>>

Converts this resource to a struct implementing [rusqlite::Params], allowing it to be passed to a rusqlite::Statement.