Trait infinity_compiler::struct_io::SqlRow

pub trait SqlRow: Sized {
    const WIDTH: usize;

    fn get_field(&self, i: usize) -> ToSqlOutput<'_>;
    fn from_row_at(row: &Row<'_>, offset: usize) -> Result<Self>;
    fn fieldtype(i: usize) -> FieldType;
    fn fieldname_ctx(ctx: &'static str, i: usize) -> &'static str;
    fn primary_index() -> Option<usize>;

    fn field_create_text(_i: usize) -> &'static str { ... }
    fn as_params(&self) -> ParamsFromIter<ContentIterator<'_, Self>> { ... }
    fn from_row(row: &Row<'_>) -> Result<Self> { ... }
    fn fieldname(i: usize) -> &'static str { ... }
    fn field(i: usize) -> Field { ... }
    fn fields() -> Vec<Field>  { ... }
    fn select_query_gen(
        db: &impl DbInterface,
        n1: impl Display,
        n2: impl Display,
        cond: impl Display
    ) -> Result<TypedStatement<'_, Self>> { ... }
}

A struct mapped to a SQL row.

Required Associated Constants

const WIDTH: usize

The number of fields in the SQL row.

Required Methods

fn get_field(&self, i: usize) -> ToSqlOutput<'_>

Returns the i-th field of the resource, wrapped in a [rusqlite::types::ToSqlOutput] for trivial insertion into a SQL statement. NOTE: rusqlite does not allow pushing Result<ToSqlOutput> values back into a [rusqlite::ParamsFromIter] iterator, so we need to unwrap those values. Since we do so from a controlled struct, this should not panic; however, it would be better to be able to propagate errors.

fn from_row_at(row: &Row<'_>, offset: usize) -> Result<Self>

Reads a whole [rusqlite::Row] into a struct, or raises a conversion error.

fn fieldtype(i: usize) -> FieldType

Returns the type of the i-th field.

fn fieldname_ctx(ctx: &'static str, i: usize) -> &'static str

Builds the name of the i-th field from context.

fn primary_index() -> Option<usize>

Provided Methods

fn field_create_text(_i: usize) -> &'static str

Any supplemental material introduced at table creation (constraints, etc.)

fn as_params(&self) -> ParamsFromIter<ContentIterator<'_, Self>>

Converts this resource to a struct implementing [rusqlite::Params], allowing it to be passed to a rusqlite::Statement.

fn from_row(row: &Row<'_>) -> Result<Self>

Reads a whole [rusqlite::Row] into a struct.

fn fieldname(i: usize) -> &'static str

Returns the name of the i-th field, as a string.

fn field(i: usize) -> Field

Return the i-th element of schema

fn fields() -> Vec<Field>

Returns the set of fields, as a vector.

fn select_query_gen(
db: &impl DbInterface,
n1: impl Display,
n2: impl Display,
cond: impl Display
) -> Result<TypedStatement<'_, Self>>

The select Statement for an object.

The ID field (if present) is not selected. (This is used only for building game files, where this field is not used anyway).

The name of the table is {n1}{n2}; since many tables are built in two parts (e.g. load_{name}, strings_{lang}), this avoids calling format! on the caller side and slightly simplifies the API.

This returns a Result<TypedStatement>: it is possible to iterate over the result and recover structures of the original type.