Trait infinity_compiler::resources::Resource

pub trait Resource: Sized {
    type Context;

    const SCHEMA: Schema<'static>;

    fn ins(&self, s: &mut Statement<'_>, key: &Self::Context) -> Result<()>;
    fn sel(r: &Row<'_>) -> Result<(Self, Self::Context)>;

    fn select_query_gen(
        db: &impl DbInterface,
        n1: impl Display,
        n2: impl Display,
        cond: impl Display
    ) -> Result<TypedStatement<'_, Self>> { ... }
    fn select_query(
        db: &impl DbInterface,
        s: impl Display
    ) -> Result<TypedStatement<'_, Self>> { ... }
}

Interface for a game resource.

This trait is implemented by the corresponding derive macro.

This is the main trait for game resources, containing the low-level interaction with the database (ins, sel). Concrete implementations are provided by the Resource derive macro.

Required Associated Types

type Context

Additional data saved in the same row as a game object; usually some identifier for the object (e.g. its resref).

Required Associated Constants

const SCHEMA: Schema<'static>

The internal description of the fields of this structure, as used to produce SQL statements.

Required Methods

fn ins(&self, s: &mut Statement<'_>, key: &Self::Context) -> Result<()>

The low-level insert function for an object to a database row.

fn sel(r: &Row<'_>) -> Result<(Self, Self::Context)>

The low-level select function from a database row to an object.

Provided Methods

fn select_query_gen(
db: &impl DbInterface,
n1: impl Display,
n2: impl Display,
cond: impl Display
) -> Result<TypedStatement<'_, Self>>

The select Statement for an object.

The ID field (if present) is not selected. (This is used only for building game files, where this field is not used anyway).

The name of the table is {n1}{n2}; since many tables are built in two parts (e.g. load_{name}, strings_{lang}), this avoids calling format! on the caller side and slightly simplifies the API.

This returns a Result<TypedStatement>: it is possible to iterate over the result and recover structures of the original type.

fn select_query(
db: &impl DbInterface,
s: impl Display
) -> Result<TypedStatement<'_, Self>>

Particular case of SELECT statement used for saving to game files.