Trait infinity_compiler::resources::Resource

pub trait Resource: SqlRow {
    fn schema() -> Schema;

    fn select_typed(
        db: &impl DbInterface,
        s: impl Display
    ) -> Result<TypedStatement<'_, Self>> { ... }
}

One of the resource tables stored in the database.

Methods attached to this trait are those which depend both on

having a full schema (with table name, etc.), and

having access to the attached (Rust) type (for typed inserts, etc.). Methods depending only on the list of columns of the schema and the Rust struct go to SqlRow. Methods depending only on the schema go to Schema. Methods depending only on the list of columns from the schema (i.e. the intersection of both cases) go to [ColumnWriter]

Required Methods

fn schema() -> Schema

Provided Methods

fn select_typed(
db: &impl DbInterface,
s: impl Display
) -> Result<TypedStatement<'_, Self>>

Particular case of SELECT statement used for saving to game files.